Use of python logging system.
Remove the previous system to save the error manually and
use python logging system instead.

Modify the name of the Raise error and use it in all the module
associated to a proper error message for the log system.

Convert if/else test to try/except to catch error to be more pythonic.