Correct a problem in the separation of the aphorisms.
Aphorisms are number from 1 and that number must be on a new line
and eventually be followed by a . and that's it.

The problem appeared because in some aphorisms there are some reference
to the line like:

Kühn_xviii 19, line 3.

The software was using that number as it was an aphorisms separation.
The bug is corrected in two different place by asking the presence of
a new line before the number:

- separation title/text (multiple 1. was detected)
- creation of the aphorisms dictionary and validation