Software is producing a TEI valid XML
Thanks to information provided by Marjorie Burghart on the TEI, the XML is now
a valid TEI one.
There are only one XML file created which simplify a lot the treatment
and diffusion of the aphorisms.
In addition to that the software is now validating automatically that
the XML file is TEI valid in respect of the schema provided by the template.
It is also possible to ask for a validation for a specfic schema using an input option.

The tests were adapted to reflect the changes.
The template is now using the last stable version of the TEI schema per default.